requirements.txt
torch
torchvision
torchaudio
numpy
pillow
scikit-learn
tqdm
gradio
